added User and Admin restrictions

+++ src/app/components/tasks/tasks.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
+import { AuthService } from 'src/app/auth/services/auth.service';
 import { TaskService } from 'src/app/services/task.service';
 import { Task } from 'src/app/Task';
 
@@ -11,7 +13,7 @@ import { Task } from 'src/app/Task';
 export class TasksComponent implements OnInit {
   tasks: Task[] = [];
 
-  constructor(private taskService: TaskService) { }
+  constructor(private authService: AuthService, private taskService: TaskService, private router: Router) { }
 
   ngOnInit(): void {
     this.taskService.getTasks().subscribe((tasks) => {
@@ -20,22 +22,38 @@ export class TasksComponent implements OnInit {
   }
 
   deleteTask(task: Task) {
-    this.taskService.deleteTask(task).subscribe(() => {
-      this.tasks = this.tasks.filter(t => t.id !== task.id);
-    });
+    if(this.authService.loggedIn()) {
+      this.taskService.deleteTask(task).subscribe(() => {
+        this.tasks = this.tasks.filter(t => t.id !== task.id);
+      });
+    } else {
+      alert("login first!");
+      this.router.navigate(["/login"]);
+    }
+    
   }
 
   toggleReminder(task: Task) {
-    task.reminder = !task.reminder;
-    //console.log(task.reminder);
-    this.taskService.updateTaskReminder(task).subscribe();
+    if(this.authService.loggedIn()) {
+      task.reminder = !task.reminder;
+      //console.log(task.reminder);
+      this.taskService.updateTaskReminder(task).subscribe();
+    } else {
+      alert("login first!");
+      this.router.navigate(["/login"]);
+    }
+   
   }
 
   addTask(task: Task) {
-    this.taskService.addTask(task).subscribe((task) => {
-      this.tasks.push(task);
-    });
+    if(this.authService.loggedIn()) {
+      this.taskService.addTask(task).subscribe((task) => {
+        this.tasks.push(task);
+      });
+    } else {
+      alert("login first!");
+      this.router.navigate(["/login"]);
+    }
+    
   }
-
-  
 }
